+++ examples/agents/adk/agent.py
@@ -1,12 +1,13 @@
 """Core module for agent orchestration"""
 # pylint: disable=protected-access
 from typing import List
+from dotenv import load_dotenv
+
 from google.adk.agents import Agent
 from google.adk.tools import ToolContext
-from adk_agent.tools import get_current_time, calculate_expression
-from adk_agent.custom_memory_tool import CustomMemoryTool
 
-from dotenv import load_dotenv
+from .tools import get_current_time, calculate_expression
+from .custom_memory_tool import CustomMemoryTool
 
 from shared.memory_client import MemoryClient
 # To load the google api keys
