+++ examples/agents/langchain/agent.py
@@ -7,10 +7,10 @@ from langchain.agents.middleware import DynamicSystemPromptMiddleware
 from langgraph.checkpoint.memory import InMemorySaver
 from langchain_google_genai import ChatGoogleGenerativeAI
 
-from langchain_agent.tools import get_current_time, calculate_expression
-from langchain_agent.custom_memory_tool import (CustomAgentState,
-                                                preload_memory_prompt,
-                                                preload_memory_prompt_for_middleware)
+from .tools import get_current_time, calculate_expression
+from .custom_memory_tool import (CustomAgentState,
+                                 preload_memory_prompt,
+                                 preload_memory_prompt_for_middleware)
 
 load_dotenv()
