+++ examples/agents/langchain/custom_memory_tool.py
@@ -4,7 +4,7 @@ from langchain_core.messages import SystemMessage
 from langchain.agents.middleware import AgentState
 from langgraph.runtime import Runtime
 
-from langchain_agent.prompt import BASE_SYSTEM_INSTRUCTIONS
+from .prompt import BASE_SYSTEM_INSTRUCTIONS
 from shared.memory_client import MemoryClient
 
 memory = MemoryClient()
