+++ examples/agents/langchain/response_manager.py
@@ -5,8 +5,8 @@ import logging
 from langchain_core.runnables import RunnableConfig
 from langchain_core.messages import message_to_dict
 
-from langchain_agent.agent import root_agent
-from langchain_agent.utils import format_messages_to_events
+from .agent import root_agent
+from .utils import format_messages_to_events
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(
